created a new list named todos.
replaced ConsoleWriteLine() with AddTodo() method to switch statement under case a, A.

 wrote void AddTodo() method.
this method takes the users inputed string and checks it to make sure its not empty and that its uniquie.
if it passes these two checks the users inputed string (var description) is added to the todos list using the todos.Add(description) method.