Replaced Console.WriteLine() with SeeAllTodos() under switch statement case's s, S.
wrote void SeeAllTodos() method.
this method counts the elements in the todos list using todos.Count

if todos.Count == 0
a message is printed to console warning user no todos have been added yet.

else a for loop is exectued that prints the index and the string for each element in the list